Fix tp zone shenanigans
